+++ src/URDFExporter.js
@@ -301,7 +301,7 @@ class URDFExporter {
 
     // Convert the object into a urdf and get the associated
     // mesh and texture data
-    parse(object, jointfunc, options = {}) {
+    parse(object, jointfunc, onComplete, options = {}) {
 
         options = Object.assign({
 
@@ -525,7 +525,11 @@ class URDFExporter {
         // format the final output
         const finalurdf = this._format(options.collapse ? this._collapseLinks(urdf) : urdf);
 
-        return { data: finalurdf, meshes, textures };
+        const result = { data: finalurdf, meshes, textures };
+
+        if (onComplete) onComplete(result);
+
+        return result;
 
     }